добавили активити для просмотра вышивки, поменяли кнопки первого экрана
