+++ app/src/main/res/layout/activity_main.xml
@@ -25,13 +25,14 @@
             app:layout_constraintHorizontal_bias="0.134" />
 
     <Button
-            android:id="@+id/open"
+            android:id="@+id/button_login_google"
             android:layout_width="120dp"
-            android:layout_height="60dp"
-            android:onClick="open"
-            android:text="@string/open_text"
+            android:layout_height="65dp"
+            android:onClick="loginGoogle"
+            android:text="@string/google"
             android:backgroundTint="@color/background"
             android:textColor="@color/black"
+            android:textSize="12sp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.173"
@@ -41,11 +42,12 @@
 
 
     <Button
-            android:id="@+id/download"
+            android:id="@+id/button_login_without_google"
             android:layout_width="120dp"
-            android:layout_height="60dp"
-            android:text="@string/download_text"
-            android:onClick="download"
+            android:layout_height="65dp"
+            android:text="@string/without_google"
+            android:onClick="loginWithoutGoogle"
+            android:textSize="12sp"
             android:backgroundTint="@color/background"
             android:textColor="@color/black"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -81,17 +83,6 @@
             app:layout_constraintHorizontal_bias="0.171"
             app:layout_constraintVertical_bias="0.849" />
 
-    <ImageView
-            android:layout_width="120dp"
-            android:layout_height="85dp"
-            app:srcCompat="@drawable/download_pic"
-            android:id="@+id/imageView3"
-            android:backgroundTint="@color/background"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintHorizontal_bias="0.893"
-            app:layout_constraintVertical_bias="0.851" />
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
