+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,6 @@
         STRING</string>
     <string name="open_text">Открыть схему</string>
     <string name="download_text">Загрузить картинку</string>
+    <string name="without_google">войти без Google аккаунта</string>
+    <string name="google">войти через Google аккаунт</string>
 </resources>
